Added functional listing of animals and their diet stats in view, added temporary security change to allow h2-console

+++ src/main/java/com/tss/repositories/AnimalReport.java
@@ -20,7 +20,7 @@ public class AnimalReport {
         this.animalDiet = animalDiet;
     }
     
-    public int getDietCount() {
+    public long getDietCount() {
         return dietCount;
     }
